initialised job ids

+++ trunk/src/org/cscs/jprinterface/lpd/QueueManager.java
@@ -9,7 +9,7 @@ public interface QueueManager {
 	
 	public List<PrintJob> getQueue(String name);
 
-	public int getNextJobId();
+	public long getNextJobId();
 
 	public Set<String> getQueues();
